feat: improve ArgoCD tab with list view, pagination, and refactoring
  - Refactor ArgoCD functionality from App.tsx to separate ArgoCD.tsx component
  - Add operation feedback for ArgoCD operations (sync, refresh, suspend/resume)
  - Implement alphabetical sorting of applications
  - Add card/list view toggle with working pagination
  - Fix list view table implementation with proper action buttons
  - Add comprehensive test coverage for ArgoCD component
  - Remove autocorrect/autocapitalize from environment profile input
  - Remove verbose debug logging from environment import
