+++ frontend/package.json.md5
@@ -1,1 +1,1 @@
-ff54a0e832d39dc500967ee6472a10b6
+ecc5e18efe679f66ef3ef4a70bb709ff
